Finished authentication implementation
