+++ config.json
@@ -1,10 +1,12 @@
 {
-	"targetOrganisation": "ahm-monash",
+	"targetOrganisation": "octokit",
 	"requireAuthentication": true,
 	"rankCutoff": {
 		"major": "0.6.-1",
 		"minor": "0.5.-1"
 	},
 	"timeUntilRefresh": 5,
-	"targetPercentage": 80
+	"targetPercentage": 80,
+	"client_id": "5cd550d6a19995e8faf0",
+	"redirect_uri": "http://localhost:3000/signin/"
 }
