+++ pages/api/loadLatest.ts
@@ -2,6 +2,7 @@ import * as fs from "fs";
 import path from "path";
 import { NextApiRequest, NextApiResponse} from 'next'
 import config from "../../config.json"
+import { checkAuthorisation } from "../../src/authenticationMiddleware";
 
 // Cache files are stored inside ./next folder
 const CachePath = path.resolve("./dynamicCache.json")
@@ -9,6 +10,11 @@ const CachePath = path.resolve("./dynamicCache.json")
 const timeUntilRefresh = config.timeUntilRefresh * 60 * 1000 // minutes to milliseconds
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+	const isAuthorised = await checkAuthorisation(req, res)
+	if(!isAuthorised){
+		return
+	}
+	
 	let cachedData = null
 
 	const current = Date.now().valueOf()
