+++ pages/api/loadNew.ts
@@ -7,6 +7,7 @@ import { NextApiRequest, NextApiResponse} from 'next'
 import { getJsonStructure } from "evergreen-org-crawler/src/index"
 
 import config from "../../config.json"
+import {checkAuthorisation} from "../../src/authenticationMiddleware";
 
 // Cache files are stored inside ./next folder
 export const CachePath = path.resolve(process.env.DYNAMIC_CACHE_PATH || "./dynamicCache.json")
@@ -69,6 +70,11 @@ export async function createData(request: "npm" | "PyPI" | "RubyGems" | null = n
 }
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+	const isAuthorised = await checkAuthorisation(req, res)
+	if(!isAuthorised){
+		return
+	}
+	
 	let cachedData = null
 
 	const current = Date.now().valueOf()
